// src/interfaces.ts
export interface productInterface {
    id: number
    title: string
    description: string
    price: number
    discountPercentage: number
    rating: number
    stock: number
    brand: string
    category: string
    thumbnail: string
    images: string[]
}

export interface categoryData {
    products: productInterface[]
    total: number
    skip: number
    limit: number
}